requirements.txt
# Core ML 
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scikit-learn==1.3.0
pillow>=10.0.0,<=10.0.1
opencv-python>=4.8.0,<=4.8.1.78

# Viz
matplotlib>=3.7.0,<=3.7.2
seaborn>=0.12.0,<=0.12.2

# MLOps
mlflow>=2.7.1,<=2.9.0
dvc>=3.20.0,<=3.44.1

# PyTorch extras
tqdm==4.65.0

# Serving M2
fastapi==0.103.0
uvicorn==0.23.2
streamlit==1.28.1
kagglehub

# Dev tools
plotly==5.16.1
pytest==7.4.1
black==23.9.1

flask==2.3.3
pydantic==2.3.0
httpx==0.25.0
pytest-cov==4.1.0
pyyaml==6.0.1

joblib==1.3.1
python-dotenv==1.0.0
requests==2.31.0
flake8==6.1.0
prometheus-client==0.17.1
python-json-logger==2.0.7
